refactor: move seat map classes to separate directory
