+++ assets/js/seat_map/seat_map_base.js
