+++ assets/js/seat_map/seat_map_editor_class.js
