+++ assets/js/seat_map/seat_map_kiosk_class.js
