+++ assets/js/seat_map/seat_map_viewer_class.js
